Rename decorators2.ts to decorators.ts
